css-text/word-break/word-break-normal-hi-000.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8">
<title>word-break: normal, hindi</title>
<meta name="assert" content="Simple Devanagari text will break on spaces under customary rules for word-break: normal.">
<link rel='help' href='https://drafts.csswg.org/css-text-3/#word-break-property'>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<style type='text/css'>
.test { word-break: normal; }
/* the CSS below is not part of the test */
.test, .ref { margin: 20px;  font: 3em/1 sans-serif; position:absolute; top: 0; width:90%;}
.ref { color:red; z-index: -100; }
</style>
</head>
<body>
<div id='instructions'>Test passes if you see no red characters.</div>
<div style="position:relative">
<div class="test" lang="hi"><div id="testdiv"><span id="testspan">हिंदी हिंदी हिंदी हिंदी</span><span id="testend">←</span></div></div>
<div class="ref" lang="hi"><span>हिंदी हिंदी हिंदी<br/>हिंदी</span><span id="refend">←</span></div>
</div>
<div id="log" style="margin-top: 10em;"></div>
<script>
var sentenceWidth = document.getElementById('testspan').offsetWidth
document.getElementById('testdiv').style.width = String(sentenceWidth - 5)+'px'
test(function() {
assert_true(document.getElementById('testend').offsetLeft === document.getElementById('refend').offsetLeft);
}, ' Hindi text should wrap at word boundaries');
</script>
</body>
</html>
